# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BloodSync API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Donor {
  id: number;
  name: string;
  blood_group: string;
  district: string;
  whatsapp_number: string;
  smoker: boolean;
  last_donation_date?: string | null;
  is_willing_to_donate: boolean;
  total_requests_received: number;
  successful_donations: number;
  created_at: string;
}

export interface UpdateDonorBody {
  is_willing_to_donate?: boolean;
  last_donation_date?: string | null;
  name?: string;
  blood_group?: string;
  district?: string;
  whatsapp_number?: string;
  smoker?: boolean;
  successful_donations?: number;
}

export interface CreateDonorBody {
  name: string;
  blood_group: string;
  district: string;
  whatsapp_number: string;
  smoker: boolean;
  last_donation_date?: string | null;
  is_willing_to_donate: boolean;
}

export type DonationRequestStatus =
  (typeof DonationRequestStatus)[keyof typeof DonationRequestStatus];

export const DonationRequestStatus = {
  pending: "pending",
  accepted: "accepted",
  completed: "completed",
  rejected: "rejected",
} as const;

export interface DonationRequest {
  id: number;
  donor_id: number;
  requester_identifier: string;
  status: DonationRequestStatus;
  created_at: string;
}

export interface CreateRequestBody {
  donor_id: number;
  requester_identifier: string;
}

export type DonationVerificationVerificationStatus =
  (typeof DonationVerificationVerificationStatus)[keyof typeof DonationVerificationVerificationStatus];

export const DonationVerificationVerificationStatus = {
  pending: "pending",
  verified: "verified",
  rejected: "rejected",
} as const;

export interface DonationVerification {
  id: number;
  donor_id: number;
  recipient_details: string;
  proof_document_url?: string | null;
  verification_status: DonationVerificationVerificationStatus;
  created_at: string;
}

export type UpdateVerificationBodyVerificationStatus =
  (typeof UpdateVerificationBodyVerificationStatus)[keyof typeof UpdateVerificationBodyVerificationStatus];

export const UpdateVerificationBodyVerificationStatus = {
  verified: "verified",
  rejected: "rejected",
} as const;

export interface UpdateVerificationBody {
  verification_status: UpdateVerificationBodyVerificationStatus;
}

export interface CreateVerificationBody {
  donor_id: number;
  recipient_details: string;
  proof_document_url?: string | null;
}

export interface StatsSummary {
  total_donors: number;
  willing_donors: number;
  total_requests: number;
  completed_donations: number;
  pending_verifications: number;
}

export interface BloodGroupStat {
  blood_group: string;
  count: number;
}

export type ListDonorsParams = {
  blood_group?: string;
  district?: string;
  is_willing_to_donate?: boolean;
};