# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BloodSync API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all donors
 */
export const ListDonorsQueryParams = zod.object({
  blood_group: zod.coerce.string().optional(),
  district: zod.coerce.string().optional(),
  is_willing_to_donate: zod.coerce.boolean().optional(),
});

export const ListDonorsResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  blood_group: zod.string(),
  district: zod.string(),
  whatsapp_number: zod.string(),
  smoker: zod.boolean(),
  last_donation_date: zod.string().nullish(),
  is_willing_to_donate: zod.boolean(),
  total_requests_received: zod.number(),
  successful_donations: zod.number(),
  created_at: zod.string(),
});
export const ListDonorsResponse = zod.array(ListDonorsResponseItem);

/**
 * @summary Register as a blood donor
 */
export const CreateDonorBody = zod.object({
  name: zod.string(),
  blood_group: zod.string(),
  district: zod.string(),
  whatsapp_number: zod.string(),
  smoker: zod.boolean(),
  last_donation_date: zod.string().nullish(),
  is_willing_to_donate: zod.boolean(),
});

/**
 * @summary Get donor by ID
 */
export const GetDonorParams = zod.object({
  id: zod.coerce.number(),
});

export const GetDonorResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  blood_group: zod.string(),
  district: zod.string(),
  whatsapp_number: zod.string(),
  smoker: zod.boolean(),
  last_donation_date: zod.string().nullish(),
  is_willing_to_donate: zod.boolean(),
  total_requests_received: zod.number(),
  successful_donations: zod.number(),
  created_at: zod.string(),
});

/**
 * @summary Update donor availability status
 */
export const UpdateDonorParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateDonorBody = zod.object({
  is_willing_to_donate: zod.boolean().optional(),
  last_donation_date: zod.string().nullish(),
  name: zod.string().optional(),
  blood_group: zod.string().optional(),
  district: zod.string().optional(),
  whatsapp_number: zod.string().optional(),
  smoker: zod.boolean().optional(),
  successful_donations: zod.number().optional(),
});

export const UpdateDonorResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  blood_group: zod.string(),
  district: zod.string(),
  whatsapp_number: zod.string(),
  smoker: zod.boolean(),
  last_donation_date: zod.string().nullish(),
  is_willing_to_donate: zod.boolean(),
  total_requests_received: zod.number(),
  successful_donations: zod.number(),
  created_at: zod.string(),
});

/**
 * @summary Delete a donor
 */
export const DeleteDonorParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary List donation requests
 */
export const ListRequestsResponseItem = zod.object({
  id: zod.number(),
  donor_id: zod.number(),
  requester_identifier: zod.string(),
  status: zod.enum(["pending", "accepted", "completed", "rejected"]),
  created_at: zod.string(),
});
export const ListRequestsResponse = zod.array(ListRequestsResponseItem);

/**
 * @summary Create a donation request
 */
export const CreateRequestBody = zod.object({
  donor_id: zod.number(),
  requester_identifier: zod.string(),
});

/**
 * @summary Approve or reject a donation verification
 */
export const UpdateVerificationParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateVerificationBody = zod.object({
  verification_status: zod.enum(["verified", "rejected"]),
});

export const UpdateVerificationResponse = zod.object({
  id: zod.number(),
  donor_id: zod.number(),
  recipient_details: zod.string(),
  proof_document_url: zod.string().nullish(),
  verification_status: zod.enum(["pending", "verified", "rejected"]),
  created_at: zod.string(),
});

/**
 * @summary List donation verifications
 */
export const ListVerificationsResponseItem = zod.object({
  id: zod.number(),
  donor_id: zod.number(),
  recipient_details: zod.string(),
  proof_document_url: zod.string().nullish(),
  verification_status: zod.enum(["pending", "verified", "rejected"]),
  created_at: zod.string(),
});
export const ListVerificationsResponse = zod.array(
  ListVerificationsResponseItem,
);

/**
 * @summary Submit donation verification
 */
export const CreateVerificationBody = zod.object({
  donor_id: zod.number(),
  recipient_details: zod.string(),
  proof_document_url: zod.string().nullish(),
});

/**
 * @summary Get BloodSync dashboard summary stats
 */
export const GetStatsSummaryResponse = zod.object({
  total_donors: zod.number(),
  willing_donors: zod.number(),
  total_requests: zod.number(),
  completed_donations: zod.number(),
  pending_verifications: zod.number(),
});

/**
 * @summary Get donor count breakdown by blood group
 */
export const GetBloodGroupStatsResponseItem = zod.object({
  blood_group: zod.string(),
  count: zod.number(),
});
export const GetBloodGroupStatsResponse = zod.array(
  GetBloodGroupStatsResponseItem,
);

/**
 * @summary Get top donors by successful donations
 */
export const GetTopDonorsResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  blood_group: zod.string(),
  district: zod.string(),
  whatsapp_number: zod.string(),
  smoker: zod.boolean(),
  last_donation_date: zod.string().nullish(),
  is_willing_to_donate: zod.boolean(),
  total_requests_received: zod.number(),
  successful_donations: zod.number(),
  created_at: zod.string(),
});
export const GetTopDonorsResponse = zod.array(GetTopDonorsResponseItem);